Break Header into multiple components
