+++ context/app/static/js/components/Header/HeaderAppBar/index.js
@@ -1,0 +1,3 @@
+import HeaderAppBar from './HeaderAppBar';
+
+export default HeaderAppBar;
