+++ context/app/static/js/components/Header/HeaderAppBar/style.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import AppBar from '@material-ui/core/AppBar';
+
+const StyledAppBar = styled(AppBar)`
+  margin-bottom: ${(props) => props.theme.spacing(2)}px;
+`;
+
+export { StyledAppBar };
