+++ context/app/static/js/components/Header/HeaderContent/index.js
@@ -1,0 +1,3 @@
+import HeaderContent from './HeaderContent';
+
+export default HeaderContent;
