+++ context/app/static/js/components/Header/index.js
@@ -1,3 +1,3 @@
 import Header from './Header';
 
-export { Header };
+export default Header;
